requirements.txt
streamlit
huggingface_hub
textwrap3
python-dotenv